chore(react): implement shared component in react and minor changes in action select
