+++ packages/react/src/components/CollectionsWrapper/Sidebar.tsx
@@ -3,6 +3,7 @@ import {
   Bars3Icon,
   XMarkIcon,
 } from "@heroicons/react/24/outline";
+import { highlightMatches } from "@honohub/shared";
 import {
   Button,
   Drawer,
@@ -16,13 +17,7 @@ import {
   useBreakpointValue,
 } from "@rafty/ui";
 import Fuse, { type RangeTuple } from "fuse.js";
-import {
-  type HTMLAttributes,
-  type ReactNode,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { type HTMLAttributes, useEffect, useMemo, useState } from "react";
 import { NavLink, useLocation } from "react-router-dom";
 import type { CollectionType } from "../../types";
 import { getPluralLabel } from "../../utils";
@@ -203,39 +198,4 @@ function SidebarItem({ link, label, matches }: SidebarItem) {
       </Text>
     </NavLink>
   );
-}
-
-function highlightMatches(
-  inputText: string,
-  regions: readonly RangeTuple[] = [],
-) {
-  const children: ReactNode[] = [];
-  let nextUnhighlightedRegionStartingIndex = 0;
-
-  regions.forEach((region, i) => {
-    const lastRegionNextIndex = region[1] + 1;
-
-    children.push(
-      ...[
-        inputText
-          .substring(nextUnhighlightedRegionStartingIndex, region[0])
-          .replace(" ", "\u00A0"),
-        <span key={`${i}-${region}`} className="bg-yellow-200/80">
-          {inputText
-            .substring(region[0], lastRegionNextIndex)
-            .replace(" ", "\u00A0")}
-        </span>,
-      ],
-    );
-
-    nextUnhighlightedRegionStartingIndex = lastRegionNextIndex;
-  });
-
-  children.push(
-    inputText
-      .substring(nextUnhighlightedRegionStartingIndex)
-      .replace(" ", "\u00A0"),
-  );
-
-  return children;
 }
